requirements.txt
# $ conda env create --file <this file>
drmaa
Django
djangorestframework
django-filter
django-extensions
django-cors-headers
environs
PyJWT
PyMySQL
pytimeparse
PyYAML
requests
mysqlclient
django-admin-rangefilter
